+++ NEWS.md
@@ -1,3 +1,19 @@
+# naomi 0.0.11
+
+* Fix issues with naomi model run options schema templates
+
+# naomi 0.0.10
+
+* Use iso3 to locate country plotting metadata instead of country name
+
+# naomi 0.0.9
+
+* Plotting metadata now returns default if country is missing from config
+
+# naomi 0.0.8
+
+* Iterate model options endpoint to be able to exclude optional data set options
+
 # naomi 0.0.7
 
 * Add function to return model run options template to build front end declaratively